move fake UserProvider to Provider namespace
