+++ tests/Functional/Fake/Bridge/Symfony/Kernel.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Functional\Fake\Bridge\Symfony;
 
-use Functional\Fake\Bridge\Symfony\Factory\UserProvider as FakeUserFactory;
+use Functional\Fake\Bridge\Symfony\Provider\UserProvider as FakeUserProvider;
 use Shippeo\Heimdall\Bridge\Symfony\Bundle\MonitoringBundle;
 use Shippeo\Heimdall\Bridge\Symfony\Bundle\Provider\UserProvider;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
@@ -47,7 +47,11 @@ final class Kernel extends \Symfony\Component\HttpKernel\Kernel
      */
     protected function configureContainer(ContainerBuilder $container, LoaderInterface $loader): void
     {
-        $container->register(UserProvider::class, FakeUserFactory::class);
+        $container
+            ->register(UserProvider::class, FakeUserProvider::class)
+            ->setPublic(true)
+        ;
+
         $container->loadFromExtension(
             'framework',
             [
